Refactor gradle build files to support multiple provider implementations
