+++ gradle.properties
@@ -1,3 +1,4 @@
 version=0.0.0
 kotlin.code.style=official
 org.gradle.jvmargs=-Xmx3G -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+org.gradle.configureondemand=true
